+++ CHANGELOG.md
@@ -18,3 +18,4 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - fix/1200713875634268 - Fix to comply with validators (getChainItems)
 - fix/1200624131677614 - Fix validator - Remove config and dumpable. Add a description
 - fix/1200203297621245 - Remove extra ending slash on found directory
+- fix/1201487027413409 - Cast LocalFileFinderAdapter::find(iterable $directories) to directory collection